Accept some actual input, fix some bugs
This isn't a very clean commit because i'm still trying to make things work.
main now reads from stdin. After each charater is read, the buffer is passed
to the spacial pooler and the output is displayed.

This exposed a few bugs, most notably the fact that columns were being sorted
in the wrong direction.

I'm still playing with the parameters to try and get a nice, stable output.

Other changes:

- Implement local inhibition

- Enable link-time-optimization

- Remove -Wconversion because gcc is dumb