Add another layer, circular input.
Add another layer to our simulation, so input goes through layer one and layer
one feeds to layer two.

Change how input is processed so that we don't "rotate" the buffer every time
a character is read, just replace the oldest character. So, from the
perspective of the cortex, the input is less like a sliding window and more
like a, well, a circular buffer.

Also represent each character by both itself and its inverse, which should help
with recognizing characters with many 0 bits.

Also:

    - Initialize synapses with random weights

    - Use a denser representation for displaying bitsets