Again, a bunch of unrelated changes trying to get the second layer to work better.
Probably the most important change is that we no longer conflate synapse
activation threshold and column activation threshold. They are two completely
different concepts.

Synapse strength ranges between 0 and 255, and the synapse
threshold is the point in that range at which a synapse is considered
connected.

The column activation threshold controls how many active synapses a
cell has to have to fire. It theoretically ranges from 0 to the maximum number
of synapses per column, and basically controls how much noise we're willing to
tolerate in our learned spacial patterns.

The latter is now computed as half the maximum number of synapse connections,
but should probably be configurable.

Other changes:

- Synapses at the edge of a layer now wrap around.

- Permit global inhibition, by setting radius = 0.

- Take the high bits of the LCRNG because oops you're supposed to do that
  because the low bits aren't very random.