Use a heap to pick the winning columns.
Since we only care about the top 2% or so columns, a full sort is overkill.
We can use a heap to implement a partial sort.

For now, qsort and heapsort coexist, with a flag to switch between them.
To enable fairer comparisons, both sorts have been made stable by having the
comparison functions fall back on comparing indices when the weights are equal.

Other changes:

- Add a parameter to specify the activation threshold of a columns.

- Define _FORTIFY_SOURCE=2 to get some overflow checking.

- Switch from -ggdb to -g, for valgrind.