Fix learning.
We were weakening synapses of _inactive_ columns when the corresponding input bit
was off, but we should only be weakening synapses of _active_ columns.

The point is to teach columns how to better recognize their patterns, and we
can only do that when the cell is active because that's when the pattern it is
supposed to be learning is (probably) present.